refactor: Facade Service 역할 변경
