--- src/main/java/subway/exception/DistanceCantNegative.java
@@ -1,14 +1,0 @@
-package subway.exception;
-
-import subway.error.exception.ErrorCode;
-import subway.error.exception.SubwayException;
-
-public class DistanceCantNegative extends SubwayException {
-
-    public static SubwayException THROW = new DistanceCantNegative();
-
-    public DistanceCantNegative() {
-        super(new ErrorCode(400, "거리는 기존 구간의 길이보다 작아야합니다."));
-    }
-
-}
