+++ src/main/java/subway/global/error/ErrorResponse.java
@@ -1,4 +1,4 @@
-package subway.error;
+package subway.global.error;
 
 public class ErrorResponse {
