+++ src/main/java/subway/presentation/dto/StationSaveRequest.java
@@ -1,4 +1,4 @@
-package subway.dto;
+package subway.presentation.dto;
 
 public class StationSaveRequest {
