+++ src/test/java/subway/dao/LineEntityDaoTest.java
@@ -9,6 +9,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.context.jdbc.Sql;
+import subway.domain.dao.LineDao;
 import subway.domain.entity.LineEntity;
 import subway.exception.LineNotFoundException;
