+++ src/test/java/subway/dao/SectionEntityDaoTest.java
@@ -9,6 +9,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.context.jdbc.Sql;
+import subway.domain.dao.LineDao;
+import subway.domain.dao.SectionDao;
+import subway.domain.dao.StationDao;
 import subway.domain.entity.LineEntity;
 import subway.domain.entity.SectionEntity;
 import subway.domain.entity.StationEntity;
