+++ src/test/java/subway/dao/StationEntityDaoTest.java
@@ -8,6 +8,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.context.jdbc.Sql;
+import subway.domain.dao.LineDao;
+import subway.domain.dao.SectionDao;
+import subway.domain.dao.StationDao;
 import subway.domain.entity.LineEntity;
 import subway.domain.entity.SectionEntity;
 import subway.domain.entity.StationEntity;
