+++ src/test/java/subway/facade/LineFacadeTest.java
@@ -1,4 +1,4 @@
-package subway.application;
+package subway.facade;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
@@ -11,11 +11,11 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import subway.domain.entity.LineEntity;
 import subway.domain.entity.SectionEntity;
-import subway.dto.LineRequest;
-import subway.dto.LineResponse;
-import subway.dto.SectionSaveRequest;
-import subway.facade.LineFacade;
-import subway.facade.SectionFacade;
+import subway.presentation.dto.LineRequest;
+import subway.presentation.dto.LineResponse;
+import subway.presentation.dto.SectionSaveRequest;
+import subway.service.LineService;
+import subway.service.SectionService;
 
 import java.util.List;
 
@@ -27,16 +27,16 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
-class LineServiceTest {
+class LineFacadeTest {
 
     @InjectMocks
-    LineService lineService;
-
-    @Mock
     LineFacade lineFacade;
 
     @Mock
-    SectionFacade sectionFacade;
+    LineService lineService;
+
+    @Mock
+    SectionService sectionService;
 
     @DisplayName("호선을 생성한다.")
     @Test
@@ -46,11 +46,11 @@ class LineServiceTest {
         final Long finalUpStationId = 1L;
         final Long finalDownStationId = 2L;
 
-        when(lineFacade.insert(LineEntity.of(request.getName(), request.getColor())))
+        when(lineService.insert(LineEntity.of(request.getName(), request.getColor())))
                 .thenReturn(1L);
 
         // when
-        Long lineId = lineService.createLine(request, finalUpStationId, finalDownStationId);
+        Long lineId = lineFacade.createLine(request, finalUpStationId, finalDownStationId);
 
         // then
         assertThat(lineId).isEqualTo(1L);
@@ -64,10 +64,10 @@ class LineServiceTest {
         SectionEntity sectionEntity = SectionEntity.of(lineId, upStationId, downStationId, distance);
 
         // when
-        lineService.registerStation(lineId, upStationId, downStationId, distance);
+        lineFacade.registerStation(lineId, upStationId, downStationId, distance);
 
         // then
-        verify(sectionFacade, times(1)).saveSection(SectionSaveRequest.of(sectionEntity));
+        verify(sectionService, times(1)).saveSection(SectionSaveRequest.of(sectionEntity));
     }
 
     @DisplayName("모든 호선을 조회한다.")
@@ -78,10 +78,10 @@ class LineServiceTest {
         LineResponse response2 = new LineResponse(2L, "2호선", "초록");
         LineResponse response3 = new LineResponse(3L, "3호선", "노랑");
         List<LineResponse> results = List.of(response1, response2, response3);
-        doReturn(results).when(lineFacade).findAll();
+        doReturn(results).when(lineService).findAll();
 
         // when
-        List<LineResponse> responses = lineService.getAll();
+        List<LineResponse> responses = lineFacade.getAll();
 
         // then
         assertAll(
@@ -100,10 +100,10 @@ class LineServiceTest {
     void getLineResponseById(Long id) {
         // given
         LineResponse response = new LineResponse(1L, "1호선", "검정");
-        doReturn(response).when(lineFacade).findById(id);
+        doReturn(response).when(lineService).findById(id);
 
         // when
-        LineResponse result = lineService.getLineResponseById(id);
+        LineResponse result = lineFacade.getLineResponseById(id);
 
         // then
         assertAll(
@@ -121,10 +121,10 @@ class LineServiceTest {
         LineRequest request = new LineRequest("2호선", "초록", 10);
 
         // when
-        lineFacade.update(id, request);
+        lineService.update(id, request);
 
         // then
-        verify(lineFacade, times(1)).update(id, request);
+        verify(lineService, times(1)).update(id, request);
     }
 
     @DisplayName("호선의 아이디를 입력받아 호선의 정보를 수정한다.")
@@ -132,10 +132,10 @@ class LineServiceTest {
     @CsvSource(value = {"1"})
     void deleteLineById(Long id) {
         // when
-        lineFacade.deleteById(id);
+        lineService.deleteById(id);
 
         // then
-        verify(lineFacade, times(1)).deleteById(id);
+        verify(lineService, times(1)).deleteById(id);
     }
 
 }
